#256
- useClientWidth перенесен в common/hooks (minor)
- common/hooks: декомпозиция модуля (patch)
