+++ src/common/hooks/use-client-width.ts
@@ -1,10 +1,10 @@
 import { useEffect, useState, RefObject, DependencyList } from 'react';
 
 /**
- * Возвращает ширину дочернего элемента.
- * @param ref Ref дочернего элемента.
+ * Возвращает актуальное значение clientWidth элемента.
+ * @param ref Ref с элементом.
  * @param deps Массив зависимостей от которых зависит пересчет.
- * @return Ширина.
+ * @return Значение свойства clientWidth.
  */
 export function useClientWidth(ref: RefObject<HTMLElement | null>, deps: DependencyList = []) {
   const [width, setWidth] = useState<number | null>(null);
